Fixed the folders of Activity 9 and 10.
