--- Taleon/TaleonActivity9/11.6 Exercises/Exercises 11.6.2 - Shapes/Circle.java
@@ -1,18 +1,0 @@
-public class Circle extends Shape{
-
-    private double radius;
-
-    public Circle(double radius) {
-        this.radius = radius;
-    }
-
-    @Override
-    public double getArea() {
-        return Math.PI*(radius*radius);
-    }
-
-    @Override
-    public String getName() {
-        return "Circle";
-    }
-}
