--- Taleon/TaleonActivity9/11.6 Exercises/Exercises 11.6.2 - Shapes/Shape.java
@@ -1,5 +1,0 @@
-public abstract class Shape
-{
-    public abstract double getArea();
-    public abstract String getName();
-}
