--- Taleon/TaleonActivity9/11.6 Exercises/Exercises 11.6.2 - Shapes/Square.java
@@ -1,19 +1,0 @@
-public class Square extends Shape{
-
-    private double length;
-    private double width;
-
-    public Square(double length, double width) {
-        this.length = length;
-        this.width = width;
-    }
-    @Override
-    public double getArea() {
-        return width * length;
-    }
-
-    @Override
-    public String getName() {
-        return "Square";
-    }
-}
