+++ Taleon/TaleonActvity10/10B - Shapes/Circle.java
@@ -1,0 +1,18 @@
+public class Circle extends Shape{
+
+    private double radius;
+
+    public Circle(double radius) {
+        this.radius = radius;
+    }
+
+    @Override
+    public double getArea() {
+        return Math.PI*(radius*radius);
+    }
+
+    @Override
+    public String getName() {
+        return "Circle";
+    }
+}
