+++ Taleon/TaleonActvity10/10B - Shapes/Shape.java
@@ -1,0 +1,5 @@
+public abstract class Shape
+{
+    public abstract double getArea();
+    public abstract String getName();
+}
