+++ Taleon/TaleonActvity10/10B - Shapes/Square.java
@@ -1,0 +1,19 @@
+public class Square extends Shape{
+
+    private double length;
+    private double width;
+
+    public Square(double length, double width) {
+        this.length = length;
+        this.width = width;
+    }
+    @Override
+    public double getArea() {
+        return width * length;
+    }
+
+    @Override
+    public String getName() {
+        return "Square";
+    }
+}
